Fix memory corruption when repairing files.
The file in question has an object defined with an object number of -1.
When we tried to add that to the list of objects it threw a (rangecheck)
error, which is reasonable as object numbers aren't signed. But instead
of moving on to look for the next object we carried on processing as
if we were looking for the end of the current object.

Breaking out of the loop instead of continuing solves the problem
properly.