Fixed shading4 bug and associated error-path memory leaks
There was a bug, where the expected number of components in the Decode
array of a mesh shading was incorrect (it should add 2 components for
a Function, not 1 component).

This caused memory leaks along the error path, which I have fixed.

The fixes some rendering issues in the test sample, but didn't change
anything else on the cluster tests.

Sample: tests_private/comparefiles/Bug694909.pdf (debugged with page 2)

This sample has more apparently unrelated memory leaks that I am still
looking at.