Fix buffer overflow reading strings
Not sure why this has not exhibited before, though the conditions are
difficult to reproduce.

When reading a string, we start with a buffer of 256 bytes. If we exhaust
the available space we allocate a new buffer, copy the content, and
free the old one.

The problem was that we only checked for buffer overflow at the end of
the loop and a few places (for example while processing a string escape
character '\') where we could execute continue without reaching the end
of the loop code block. If that happened to co-incide with using the
last byte of the buffer then we could overrun by one byte.

This commit moves the buffer checking to the beginning of the loop so
that we cannot avoid checking it before reading a byte.