graphics library- clean up streams used as DataSource for functions
The graphics library doesn't close streams which are supplied as
the DataSource for type 0 functions, leading to a memory leak in pdfi
because we create memory streams (and hence allocate buffers) for this
kind of input.

Add code to close the stream (if present) when the type 0 function is
freed. There may be future work required here and in other function
types.