Bug 702611: Fix tiff_from_filep to handle subclassing
tiff_from_filep() creates a tiffio structure, which contains a pointer to the
tiff device. If the device has been subclassed, however, that is the transient
copy of the tiff device, and not the memory used by the "real" device, as
pointed to by the graphics state.

Since we only use the device to get to the gs_memory_t pointer, just store that
instead.