Enhance First/LastPage processing
Previously the FirstPage and LastPage processing device did not allow
any means to reset the PageCount. This was because Ghostscript's
command line processing does not permit changing non-PostScript
controls (interpreter and some device parameters) after the first file
has been run.

GPDL however, has a new mechanism 'set_param' which can be used
programmatically, and that does permit for device and interpreter
parameters to be altered after the initial file has been processed.

To allow for this the gdevflp device now processes parameters itself
instead of relying on the underlying device to do so. The parameters
FirstPage, LastPage, PageList and DisablePageHandler now all reset the
page count to 0 when they are encountered. This means that, using gpdl,
it is possible to select a set of pages from one file, then select a
different set of pages from a second file. Sending any of these
parameters (except, obviously DisablePageHandler) also now automatically
enables the device again ie it sets DisablePageHandler to false.

It is not, unfortunately, possible to load the gdevflp device at any
time except when the underlying device is initially opened. This means
that if any file is to be processed using gdevflp the first file must
use one of the parameters, in order to load gdevflp. The simplest
solution is simply to set -dFirstPage=1 which will load the device and
run all the pages from the file.

This commit also includes a minor change to the PDF interpreter. Because
the PDF interpreter (currently) handles subsets of pages itself, it
does not want the first/last page device to be active, so it
disables the device by sending a 'DisablePageHandler' to it. However
(because of the GS command line, as described in the first paragraph) it
did not bother to re-enable the device. So here we add a line to
re-enable the device after processing is complete.

This is probably superfluous now that sending the params will re-enable
the gdevflp device anyway, but it should make the intention plain.