Fix CMap memory leaks
We were only cleaning up CMap object contents if the CMapType was 1
(ToUnicode CMaps being type 2). Previously we had initialised the
CMapType to 1, but it turns out that its perfectly legal to have a CMap
with a type of 0, and we have files where this is true.

For now clean up any CMap with a type which is not 2. Add a comment
to the code describing this, Chris plans to address this more
comprehensively when adding ToUnicode support.