Fix type 5 halftones
Type 5 halftones were fundamentally broken. Its less than obvious how
this works in the graphics library but here's a summary:

The device halftone has an 'order' structure, and in that we store the
bit_data and levels for the Default halftone. The Default is used for
any colorant not assigned a unique halftone in the type 5 dictionary.

We also (may) have a components entry in the device halftone. This is
present if the type 5 halftone has colorants which are used by the
device *other* than the Default. Again we store the bit_data and levels
in the component corder member component number in the comonent.

Previously we were trying to store all the halftones in the device
halftone order, which meant that only the last one was actually used,
and applied to all the components in the output. In addition we leaked
the other halftones because they were not referenced from the device
halftone and so were never freed.