Bug 702744: Add windows utf-8 handling for unlink and rename.
To do this, we introduce gp_unlink and gp_rename, and call those
as appropriate.

Also, make gp_open_scratch_file add the file to the permit list.
When such a file is closed, it will be deleted from the permit list.

It will also be removed from the permit list if the file is deleted
using the PS deletefile operator.

On closedown, if scratch files haven't been deleted, then we'll
delete them as part of the closedown of gs_lib_ctx.

This means that 'purging' the control lists must not remove scratch
file paths from the list.

Also, ensure that gsapi callers can't maliciously (or accidentally)
remove scratch path paths from the list so as to leave them around
after closedown.