Free CMap objects, no matter what their type
CMaps should (for the /Encoding of a type 0 font) have a CMapType of 0
or 1. ToUnicode CMaps should have a CMapType of 2.

We have a couple of files which are plainly invalid since they have a
Type 0 Font with an Encoding where the CMap has CMapType of 2. I'm not
certain if Ghostscript processes this 'correctly' but it was causing
our current code to leak the CMap.

This solves the leak, there's a message for Chris in the comments when
he comes to do ToUnicode CMaps.