Rejig internals of gs_pattern1_make_pattern.
Previously we called "compute_inst_matrix" once at the start, then
maybe either called it again (if we ADJUST_SCALE_BY_GS_TRADITION)
or manually did part of it (retransforming the bbox) (if we
ADJUST_AS_ADOBE).

Due to the vagaries of floating point, we can't be sure that
retransforming the bbox will give exactly the same result (as
our attempt to move the bbox onto a pixel origin may result
in future transformations may give x +/- epsilon rather x,
for example). Consequently, when we retransform, follow the
same procedure as we did before.

This produces some diffs. We choose to accept these (minor) diffs
to give us consistent results.