Bug702727 --  Fix rc counting of penum->pcs for image1 and image4
This relates to the structure gx_image_enum_s which is used by image1
and image4.

The fix is unfortunately scattered around the code, but I think I found
appropriate places.  It is ugly because type1 and type4 images
apparently share a lot of the same code, so the module layers are not
as they should be.

Need to increment and decrement the rc since the enum struct wants to
keep it around (especially in gs where it literally goes back out into
the gs interpreter between when the enum is built and when it is
used!).

This is part of cleaning up memory leaks for non-GC interpreters
(pdfi).

Changed the pcs entry in the enum struct to not be const, since it
needs to be reference counted.

Cleaned up an unrelated compiler warning in gxidata.c
There are also some unrelated whitespace changes.