Fix PDF /Producer for new patch level
Bug #702772 "Strange /Producer attribute"

The change to include a patch level for Ghostscript in the gs_revision
value wasn't being reflected in the code to generate the version
number used in the PDF /Producer string.

Copied the code from printf_program_ident() in base/gsmisc.c to the
code in pdf_store_default_Producer() and commented both pieces of code
to note the existence of each other so that in future hopefully we won't
forget to keep them in sync.

Thanks to Peter Cherepanov for spotting the flaw.