PDF interpreter - reject non-dictionary /Annots array entries
Bug #702745 "Error: /typecheck in --runpdf--"

The PDF file has an invalid /Annots array entry:

5 0 obj
<</Annots[5 0 41 0 R]

The entries should be dictionaries or indirect references to
dictionaries, integers aren't valid. This causes a typecheck error when
trying to read keys from the dictionary.

This commit checks that the entry is a dictionary before trying to
process it; if it isn't we raise a warning and ignore it.