txtwrite - improve CIDFont handling
The calculation of the font size for modes 0 & 1 was incorrect when the
font was a CIDFont. The size calculation relied upon the FontMatrix of
the type 0 font and did not consider the descendant.

For PDF files we can ignore the type 0 font and simply use the
descendant font matrix. For PostScript input we will probably have to
be more careful so this is an interim fix for PDF input only.