Fix gs_main_set_language_param to escape encapsulation.
If we are running in JOBSERVER mode, then postscript can't set stuff
into systemdict while in a job. We need to exit the encapsulation,
set the param, then reenter.

(Contrary to my memory, dnit_job does not exit the encapsulation - or
rather it does, but it unsets the device, and then immediately reenters
encapsulation again.)

Thanks are due to Ray for identifying the problem, and to Chris
for being a sounding board.