pdfwrite - fix numerous font-related memory leaks
The pdfwrite device was originally written to do all its memory
management via garbage collection, Unfortunately this leads to numerous
memory leaks when using an interpreter which doesn't implement a
garbage collecting memory model. Currently that's everything except
the PostScript interpreter.

Many of these leaks cannot be exercised by anything except the
PostScript interpreter (and the new pdfi interpreter) but I think some
certainly could be.

Don't increment the 'subst' structure when copying a TrueType CIDFont,
this leads to the structure never being freed.

The code to free the 'subst' CID when freeing a TrueType CIDFont was
incorrect, leading to memory crashes (previously it was simply never
executed, and leaked).

When increasing the size of the CIDMap entry for a TrueType CIDFont,
free the memory used by the original map after copying the data and
before overwriting the stored pointer.

When freeing a copied font, if the font is a TrueType CIDFont decrement
the 'subst_CID_on_WMode' member so that it gets freed.

When freeing a copied font, use the copied glyph data 'data' member
rather than the 'size' member to decide if we should free the data.
Sometimes we allocate memory for a size of 0.

When freeing a copied font, if we created a names array for the glyphs
we need to free the string copies of the names we created as well.

When freeing a copied font, if the CIDFMap entry has been created we
must free that too.

When dropping a 'complete' font copy because we won't be using it, do
not simply set the pointer to NULL but actually call code to free the
copied font properly.

When freeing a pdfwrite copy of a base font, if the CIDSet member has
been allocated we need to free it.

When adding to the pdfwrite font cache, if the glyph_usage and
real_widths members have already been allocated we must free them before
assigning newly allocated memory.