Bug 702761: Improve handline of PS file with many large ReusableStreams
This file, with its original, non-optimal, method collected 1,000 ReusableStream
files, each of which was 7.5Mb, stored as 1,000 arrays of 180 40,000 byte strings.
As the file progressed, the load on the GC (garbage collector) to mark-and-sweep
all of those strings got slower and slower.

This change collects the strings in the array of strings into a single large
string (up to max_string_size, currently 16Mb), so the load on the GC is cut
down to finding and freeing the array-of-strings since the last GC, and the
(at most) 1,000 strings, one for each ReusableStreamDecode "file" object.

The submitter has already accepted the change to avoid storing the expanded
raw Image data, but there are other usages of ReusableStreamDecode that may
be used with large files (e.g. streamed PDF input files) that can benefit
from not loading down the GC.