pdfwrite - fix another memory leak
When we close down pdfwrite we check the stack of saved gstates to
see if there are any dash patterns to be freed. We must check the
entire potential stack, not just to the current level, in case we
allocated a dash pattern copy at a deeper level than we were at when
we exited.

Because dash_patttern is allocated in non-gc memory this leak affects
all interpreters.