Bug 702806: Performance decreases throughout file, PS to PDF
A sample file of 1,000 pages from the customer allowed us to reproduce
the issue. The file creates MANY fonts (3,398 over the 1,000 pages),
and the font_cache list (not really a cache) would have the newest
created fonts at the head of the list, but later pages would use
fonts created earlier that were at the end of the list.

Fix this by moving the found font to the head of the list (MRU).
On this file, the time on a fast machine went from 152 to 65 seconds
and the times for pages does not show the growth seen originally.