XPS interpreter:  Have interpreter handle page range processing
This has the obvious benefit that the interpreter will skip processing
pages except those that lie in the FirstPage LastPage range.
If the PageList is used, the XPS interpreter will create a new
list of linked pages to process.  The XPS interpreter handles all
the PageList cases currently implemented by the PDF interpreter
including even and odd.  In addition, the XPS interpreter will handle
backward indexing (e.g. 100-1), as well as starting from last page
to another page (e.g. -2   which means from last page to second page),
and page repeats  (e.g. 1,2,1,2,1,2   or 1-3,3-1  or  1-,-1 which is
do page 1 to end and end to page 1)  Also, setting -dFirstPage=4
-dLastPage=1 will create pages 4,3,2,1