pdfwrite - fix memory leak when allocation of a base font fails
When the allocation of a PDF base font (pbfont) fails we were only
freeing the memory used by the structure, not any of the components
stored in it.

There are quite a few subsidiary elements which need to be free and
there is already code to handle this, but its bundled up in
pdf_font_descripto_free(). So to make this easier, break out the code
to free the base font from pdf_fotn_descriptor_free() into its own
routine, pdf_base_font_free() and call that from both
pdf_font_descriptor_free() and pdf_base_font_alloc().