pdfwrite - fix several memory leaks with colour spaces
The colour space creation code uses pdf_string_to_cos_name() quite
frequently, and it often then passes the resulting cos_value_t to
functions such as cos_array_put and cos_dict_put.

The basic 'put' functions actually copy scalar objects when they are
put in an array or dictionary (but they don't copy other composite
objects!) so we need to free the object created by
pdf_string_to_cos_name() after we've 'put' it.

The only function to do that is cos_value_free() but that (unreasonably)
requires a cos_object_t parameter, even though the cos_value may not
be in a cos_object. Since the only use for the cos_object is to
retrieve a memory_t pointer in order to free the data used, we can
simply replace the cos_object_t * with a gs_memory_t * and use that
directly.

This meant a few places needed to change where cos_value_free() was
already being used.

After that I rationalised the code in gdevpdfc.c to some extent. Instead
of the existing mixture of *_put() and *_put_no_copy() calls I've chosen
to use *_put() calls throughout and then explicitly free the data which
has supposedly been copied.

We could have used _nop_copy() throughout but that would have meant
checking the return value and freeing the original data if the copy
failed. It seems simpler just to unconditionally free the original
data.

In one place we also need to check that the original data was not NULL
and that it was a scalar type. This is because (as noted above) if it
isn't a scalar type the *_put() functions *don't* copy the value, so we
don't want to free the original data in this case as it has been
transferred to the cos_object.

Nancy did most of the original work on discovering the actual memory leak
here, and suggesting the basic fix, so many thanks!