pdfwrite - fix memory leak in copied font
This really should have been fixed by an earlier commit, but it had a
dumb error. We were freeing cfdata->names after we'd checked the
first entry in the table, instead of checking *every* entry in the
table first. In fact the memory was already freed blow so we didn't
need this at all.

In addition the code to free extra_names was inside the loop and didn't
need to be as its a linked list not an array. So move it outside where
it will only execute once. Should be a tiny bit faster.