Fix pdfwrite memory leak in pattern clists
This is for non-GC interpreters only.

This code is very convoluted, but I think the ->retained flag can be
used to see if it's okay to safely free the memory.

In the case of patterns, for at least some cases the ctile wasn't
in the cache so it didn't free anything in gx_pattern_cache_free_entry().