Fix memory leak with patterns
If we encountered an error while setting a shading pattern we did not
free the pattern context, and its contents.

In order to make freeing the contents simpler, break the code for doing
this out of pdfi_pattern_cleanup() and make a new function just to
free the context and its contents. Use this from pdfi_pattern_cleanup
as well.