Fix crash with copied TrueType fonts and pdfwrite
When freeing a copied font we now test for the data pointer being
non-NULL instead of checking the size of the string. This was because
we would still create strings for glyphs with a length of 0 and using
the size meant we would not free those strings, leading to a memory
leak.

However this causes problems (only for 0 length data) with TrueType
fonts regarding ownership. We avoid this by not allocating any memory
for 0 length string data.

We do need to fill in the remaining elements of the copied glyph though.