oss-fuzz 23113: Bounds check buffer in gs_scan_token()
To handle malformed/multiply negated numbers (e.g. "---1") in PDF files,
the number scanning code skips over multiple leading minus signs (*only* for
PDFs), but code skipping the minus signs did not bounds check the buffer, so
could rattle off the end.

Unfortunately, where this happens, there is no way for us to differentiate
between a (malformed) token that straddles a buffer boundary, or a (strange!)
token that is the entire buffer.

As the code in question is only active for PDF interpreting, and the Postscript
PDF interpreter uses the "token" operator to tokenize stream contents, we can
assume (in this case) the buffer is the entire token.

So, in the event we hit the end of the buffer without reaching something other
than a '-' then reset the buffer pointer, and "retry" it as a name.