Fix silly DecodeParms for compressed xref
The 'optimised' file for Bug 691220 has a compressed XRef where the
stream is Flat compressed, and has DecodeParms meaning we need a
Predictor.

The dictionary also includes a /Colors key, which is nonsensical for a
stream which is not an image. If we honour this key we end up trying
to read at least 5 colours at a time, when in reality we want to be
reading byte-wise.

The simplest solution is to remove the /Colors entry from the
DecodeParms dictionary which means we fall back to the sensible default
value of 1.

To do this, add a new dictionary function to delete key/value pairs
from dictionaries.