Increase the size of the stack
A very few mad PDF files were causing a stackoverflow error, because
they contain dictionaries with hundreds of thousands of entries. These
dictionaries are created on the stack which has a compile-time limit,
previously 32,768 entries, and so caused a stack overflow.

We were ignoring the stack overflow error (as with most errors) and
continuing which for one file with a massive Resources dictionary in
the Page object meant we simply tried to read the dictionary again for
the next object. Obviously this was extremely slow.

Increase the stack to 512K entries which is enough for all the files
in our test suite. If we come across a worse example we should move to
a dynamic maximum and allow it to be specified on the command line.

Also added some exemptions so that we drop out earlier if we hit a stack
overflow error, so that we don't spend lots of time failing to load a
dictionary, repeatedly.