pdfwrite - add a new method to detect pdfmark-capable devices
The existing method to determine if the current device is capable of
processing pdfmark operations only works in PostScript. Since pdfmark
is a PostScript operator this isn't unreasonable, but we need a way for
the pdfi interpreter to know that a device is capable of handling
pdfmarks, because some information from the current interpreter is
passed that way (eg DOCINFO).

Add a new parameter "PdfmarkCapable" which can be retrieved via a
spec_op. Any device capable of handling pdfmarks (via put_param calls)
must return true for this parameter.

Update the current PDF interpreter to use this instead of the PostScript
method.

Note: the ps2write device is *not* PdfmarkCapable.