Fix memory leak on files with no 'startxref'
If ctx->startxref was null, it was adding it to the loop detector, but
the 'mark' in a loop detector is also null, so this meant the cleartomark()
didn't clear everything because it thought the startxref was the mark.

Fixed to only add the startxref object to the loop_detector if it's
not null.

Also reworked the error handling and cleanup to funnel down to the exit.

Samples:
tests_private/comparefiles/Bug691894.pdf
tests_private/comparefiles/Bug694303.pdf