Fix several problems exhibited by file Bug692675.pdf
Initially this was a memory leak due to us not closing a partially
completed filter chain when one of the filters errored.

To fix this I need to use the code already existing in pdfi_close_file
but without having a pdf_file object. So I broke out the actual code
to close a filter chain (up to a target filter) from that routine into
its own code. If we get an error while we are building a filter chain
we now call that function to close and free the filters.

However, while investigating this I noticed that the existing code does
not throw an error on the broken DecodeParms entry (instead of a
dictionary it has an empty hex string '<>'). Instead the code carries
on without applying any DecodeParms.

So modified the code to not throw an error, but to record yet another
type of warning and emit that on exit.