Font copying - fix memory leak
When copying a glyph for a copied font, if we encountered the same glyph
twice, with the same name, we simply replaced the name in the name table
of the copied font. If there was already an entry then it would leak.

Note that a second instance of the glyph with a different name would
end up in the extra names table and doesn't leak.

Fix the leak by freeing the existing string copy if we have a new one.