Fix memory corruption with compressed objects
After creating an object on the stack from a compressed stream, we
check the newly created object to see if its object number is correct.

If it isn't, then we popped it off the stack, which frees it, but we
left the assignment to compressed_object in place. On exit we the
decremented the reference count, leading to memory corruption.

Jiggle the code so that we don't assign the object to compressed_object
until after we've validated it.