Fix commit 066cb5c8e0f6bbe7319c9861e4166705cd59dc11
Commit 066cb5c8e0f6bbe7319c9861e4166705cd59dc11 wasn't quite what I'd
intended. There's nothing wrong with it, and it does fix the problem,
but I prefer the simpler code around line 694 here.

I also noticed that we have defined a new error which isn't really an
error. This is pretty clearly a case where we should be using a > 0
return code to indicate the caller need to take action (in this case
process a fixed keyword). So I've removed the error code and altered
the functions involved to return a value >0 when we repair a keyword.
I've also #defined the value in case we want more later.

And then restore the newlines in (if code < 0) return code;