Fix memory leak when aborting CMap reading.
If we got an error reading a CMap we were not freeing the data already
allocated and filled in in the pdf_cmap object.

There is code to do this in pdfi_free_cmap, but that also frees the
cmap object itself which, in this case, we don't want to do as in this
case it is a stack-allocated object.

So break out the content freeing code to a separate routine and call
that when we need to free the cmap contents.