Fix memory leak with clip device and image width/height = 0
When rendering masked images we can create a clip device in order to do
the actual masking before passing the result to the final device.

After creating the device we store it in a 'gs_image_common_enum' for
later use. We then proceed to initialise the main image enumerator in
gs_image_common_init().

As an optimisation, if the image width or height is 0, we don't waste
any time processing it but simply end the image and exit.

However, if we had created a clip device, we did not free that device
leading to a memory leak.

This commit simply 'un-retains' the device causing it to be freed.