Bug 702871: Fix GraphicsAlphaBits + Fill/Stroke issue.
When we flush the alphabuffer, mapped_start and mapped_height are
both set to zero. When we refill it, mapped_height is set to
non-zero. USUALLY mapped_start is set to non-zero too, but it can
be set to zero when we are close to the bottom of the page.

Hence we should test for mapped_height != 0, rather than both
mapped_height and mapped_start being non-zero.

So the 'quick hack' that Chris used is actually correct.