Correct fix for stack smashing.
Commit 4d9fa68df "Fix stack smashing in blending code." was the
wrong fix (and indeed didn't work in the weekly tests).

We therefore back that out, and add the correct fix.

When using a temporary array to allow us to Hue blend, we need
that array to be large enough to cope with all the components
we might have, including spots.