pdfwrite - fix image downsampling when the input is a 16-bit image
Bug #702868 "Segmentation fault when using dPDFSETTINGS"

The problem is to do with image downsampling, and 16 bit images. When
the input image has 16 bits per component, and is sufficiently high
resolution on output that it triggers image downsampling, in this case
the image downsampling setup can't deal with the input. In addition we
*already* deal with 16-bit input by explicitly adding a filter to
reduce the input from 16 bits to 8, so its not necessary for the
downsampling code to do this.

So fix it simply by altering the colour depth form 16 to 8 in the
params we pass to the setup code.