Bug 702897: Fix shared libtiff/libjpeg check
For various reasons we cannot combine shared and not shared libjpeg and
libtiff - they either both need to be "local" or both shared.

But the check for that compatibility was triggered during the recursive
configure call when setting up for cross compiling.

Skip the check for that recursive configure call.