Fix crashes with pdfwrite, TrueType input fonts and Windows 64-bit build
With the change to 64-bits unconditionally for gs_glyph we can now
define GS_MIN_CID_GLYPH in 64-bit terms.

Previously we were using the architecture size of a long_long to
determine which define to use, and we do not define long_long on Windows
leading to us using an essentially 32-bit definition. This caused
indexing off the end of an array in copied_glyph_slot()