Bug 702904: Another fix to check for shared libjpeg/libtiff
We cannot combine shared and not shared libjpeg and libtiff - they either both
need to be "local" or both shared, and configure checks that and fails when
the two are incompatible.

However, that check would fail when either libjpeg or libtiff were not being
included at all. Since it is libtiff that is the "problem" for this
compatibility, now check if TIFF is included, and if not, skip the check.