PDF interpreter with pdfwrite fix non-square scaling stroked text
Bug #702900 "pdfwrite device causes bold text output"

The PDF interpreter has special code for dealing with line widths and
stroked text, this is because we do not preserve the CTM in the input
through to the output in pdfwrite.

What we need to do is figure out what line width with an identity CTM
would draw a line the same width as the current line width and CTM.
Normally pdfwrite takes care of this itself, but until relatively
recently it did not have access to the textmatrix, only the CTM. But for
text, by the time we get to the point of emitting the text,, the CTM
already includes the Text matrix (Tm).

So the PDF interpreter (which does have the CTM before the TM is
applied) calculates a line width and sets that before sending the text.

However, when the scaling was non-square we were not doing that, leading
pdfwrite to emit the unscaled linewidth. Intriguingly we have no files
in the test suite for which this is a problem, all the files where the
CTM is non-square it is also close to the identity so not visibly
incorrect.

The file here, however, has a non-identity CTM which is slightly
non-square when the text is rendered, and very far from the identity.

Since we can't tell which scaling to use for the linewidth we 'fudge'
it by using the transformed x and y values of the linewidth, as two
sides of a triangle, then calculating the hypotenuse. Obviously this
won't work for seriously skewed matrices but it seems unlikely we
could calculate a reasonable linewidth for that anyway.