PDF interpreter & pdfwrite - better detection of per-page output
Bug #702791 "%d vs. %03d and pdfmark destination page ... points beyond the last ..."

The PostScript code in the PDF interpreter which looks for
file-per-page output, in order to drop Outlines and Dests was
only checking for %d not variants liek %03d. This could lead
to warnings about the Destination being dropped.

The code provided by Peter checks for an odd number of '%'
characters instead. This isn't the same test as pdfwrite uses
but that is in C and not immediately available in PostScript.

Since we plan eventually to drop the PostScript implementation
of the PDF interpreter and move to a C version, this is fine
for the interim.