XPSWrite: Fix memory leaks and page resource relations
The page resource relations information tells which of the
resources are used for each page.  The information was not
getting written out correctly for the multi-page case.  The
Microsoft viewer is very fussy about this and would not
display some pages, even though the item is in the resources.

Made sure that each page writes out the proper relations to the
resources and that no duplicate relations are written.

Also fixed memory leaks from the zip file information not getting
freed as well as some TIFF client information.