pdfwrite - remove erroneous free from cos_dict_put_copy
Nancy discovered this one while working on annotations for pdfi. The
code here would attempt to free the memory pointing to the key name
which is passed as an argument to the function, if the *dictionary*
properties were set to free dictionary keys.

Obviously the function argument is not the same thing as the key stored
in the dictionary, so even at first inspection this looks incorrect.

The key is supplied (much higher up) from a param list, because the pdfi
code uses C param lists, where the data is stored in a single large
allocation, it is not possible to free an individual key and causes a
crash.

Ghostscript gets away with this for two reasons; firstly it uses a
stack param list (which is basically a ref param list) where each key
is stored in its own allocation. Secondly the key is not actually
allocated, its an entry in the interpreter name table. So when the
code in cos_dict_put_copy 'frees' the memory, nothing actually happens,
the memory manager just notes that the memory is unused now. This
avoids double freeing the memory which would otherwise happen.

So essentially this code is wrong, and we've been getting away with by
sheer luck for decades.

I tested the modification extensively, including memory leak checks with
many files run through pdfwrite and can find no problems with simply
removing the line of code.